Update maximum_resiliency.md
The direct connect Maximum resiliency design at https://docs.aws.amazon.com/directconnect/latest/UserGuide/maximum_resiliency.html shows both connections from a customer data center going to the same Direct Connect location.
If there is a failure of a DX location it would isolate a customer DC and force the traffic via the customer data center interconnect links (customer dc to customer dc) before going to AWS. 
This proposed change on the documentation is to clarify that the there are no restrictions to connect from a single customer data center to multiple DX locations. In fact, this may be desirable.
This proposed update to the documentation was discussed with Steve Seymour as well.